+++ packages/react/src/hooks/watched/useWatchedQuery.ts
@@ -16,13 +16,6 @@ export const useWatchedQuery = <RowType = unknown>(
 ): QueryResult<RowType> | ReadonlyQueryResult<RowType> => {
   const { query, powerSync, queryChanged, options: hookOptions, active } = options;
 
-  // This ref is used to protect against cases where `queryChanged` changes multiple times too quickly to be
-  // picked up by the useEffect below. This typically happens when React.StrictMode is enabled.
-  const queryChangeRef = React.useRef(false);
-  if (queryChanged && !queryChangeRef.current) {
-    queryChangeRef.current = true;
-  }
-
   function createWatchedQuery() {
     if (!active) {
       return null;
@@ -42,24 +35,55 @@ export const useWatchedQuery = <RowType = unknown>(
   }
 
   const [watchedQuery, setWatchedQuery] = React.useState(createWatchedQuery);
+  const disposePendingUpdateListener = React.useRef<() => void | null>(null);
 
   React.useEffect(() => {
     watchedQuery?.close();
-    setWatchedQuery(createWatchedQuery);
+    const newQuery = createWatchedQuery();
+    setWatchedQuery(newQuery);
+
+    return () => {
+      disposePendingUpdateListener.current?.();
+      newQuery?.close();
+    };
   }, [powerSync, active]);
 
-  // Indicates that the query will be re-fetched due to a change in the query.
-  // Used when `isFetching` hasn't been set to true yet due to React execution.
-  React.useEffect(() => {
-    if (queryChangeRef.current) {
-      watchedQuery?.updateSettings({
-        query,
-        throttleMs: hookOptions.throttleMs,
-        reportFetching: hookOptions.reportFetching
-      });
-      queryChangeRef.current = false;
-    }
-  }, [queryChangeRef.current]);
+  /**
+   * Indicates that the query will be re-fetched due to a change in the query.
+   * We execute this in-line (not using an effect) since effects are delayed till after the hook returns.
+   * The `queryChanged` value should only be true for a single render.
+   * The `updateSettings` method is asynchronous, thus it will update the state asynchronously.
+   * In the React hooks we'd like to report that we are fetching the data for an updated query
+   * as soon as the query has been updated. This prevents a result flow where e.g. the hook:
+   *  - already returned a result: isLoading, isFetching are both false
+   *  - the query is updated, but the state is still isFetching=false from the previous state
+   * We override the isFetching status while the `updateSettings` method is running (if we report `isFetching`),
+   * we override this value just until the `updateSettings` method itself will update the `isFetching` status.
+   * We achieve this by registering a `settingsWillUpdate` listener on the `WatchedQuery`. This will fire
+   * just before the `isFetching` status is updated.
+   */
+  if (queryChanged) {
+    // Keep track of this pending operation
+    watchedQuery?.updateSettings({
+      query,
+      throttleMs: hookOptions.throttleMs,
+      reportFetching: hookOptions.reportFetching
+    });
+    // This could have been called multiple times, clear any old listeners.
+    disposePendingUpdateListener.current?.();
+    disposePendingUpdateListener.current = watchedQuery?.registerListener({
+      settingsWillUpdate: () => {
+        // We'll use the fact that we have a listener at all as an indication
+        disposePendingUpdateListener.current?.();
+        disposePendingUpdateListener.current = null;
+      }
+    });
+  }
 
-  return useNullableWatchedQuerySubscription(watchedQuery);
+  const shouldReportCurrentlyFetching = (hookOptions.reportFetching ?? true) && !!disposePendingUpdateListener.current;
+  const result = useNullableWatchedQuerySubscription(watchedQuery);
+  return {
+    ...result,
+    isFetching: result?.isFetching || shouldReportCurrentlyFetching
+  };
 };